requirements.txt
snowflake-connector-python==3.12.3
python-dotenv==1.0.1
